refactor: rename Logo to GewisLogo
