+++ apps/dashboard/src/components/GewisLogo.vue
